+++ src/test/neural_network_test.rs
@@ -11,6 +11,7 @@ mod conv2d_test;
 mod flatten_test;
 mod global_average_pooling_1d_test;
 mod global_average_pooling_2d_test;
+mod global_average_pooling_3d_test;
 mod global_max_pooling_1d_test;
 mod global_max_pooling_2d_test;
 mod global_max_pooling_3d_test;